Add constraints area
Constraints are now displayed in a new area below the graph
representation of the model. The functionality to add, edit and delete
constraints still has to be implemented. For now, Unicode icons are used
as it is difficult to include images in treeview cells. But icons are
already provided in the resources folder for later integration.